pyproject.toml
[tool.poetry]
name = "fastauth"
version = "0.1.0"
description = "Authentication middleware for FastAPI with JWT and role-based access control"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{ include = "fastauth" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.6"
pydantic = "^2.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
ruff = "^0.1.3"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"

[tool.black]
line-length = 120
target-version = ["py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
line-length = 120
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
    "W",   # pycodestyle warnings
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # Do not perform function call in argument defaults (needed for FastAPI's Depends)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"] # unused imports in __init__ files

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
